Add files via upload
For this project, we worked with a company called Artemis Financial, and we worked on vulnerability assessments to ensure their data and information were secure and not vulnerable to attacks. Looking for security vulnerabilities was more complicated than I thought, mainly because my system would not allow me to complete a Self-signed certificate. I was using MAC, and it wouldn't allow me to assign the certificate to the necessary home space. We used Maven to do security checks and suppress vulnerabilities that did not need to be adjusted or changed. We rewrote code and ensured all systems used were up to date and not using older versions, which are more likely to be attacked and leave information insecure. After refracting the code, we went back in and checked for vulnerabilities again.